Ported the remaining game objects and management systems Temporarily patched code to prevent one-time animations from causing crashes

+++ Iteration 4 (Cocos2d-x)/Asteroids/Classes/SpriteAnimationSystem/SpriteAnimationCollection.cpp
@@ -33,7 +33,7 @@ SpriteAnimationCollection::~SpriteAnimationCollection() {
 }
 
 int SpriteAnimationCollection::numberOfAnimations() const {
-	return m_animations.size();
+	return static_cast<int>(m_animations.size());
 }
 
 bool SpriteAnimationCollection::hasAnimation(const SpriteAnimation & animation) const {
